+++ gui/node_gui_builder.cc
@@ -21,12 +21,10 @@ Builder::~Builder () {
 void Builder::Init (Handle<v8::Object> target) {
     HandleScope scope;
 
-    Local<FunctionTemplate> t = FunctionTemplate::New (New);
-    constructor_template = Persistent<FunctionTemplate>::New(t);
-    constructor_template->InstanceTemplate()->SetInternalFieldCount(1);
-    constructor_template->SetClassName(String::NewSymbol("Builder"));
+    CREATE_NODE_CONSTRUCTOR_INHERIT ("Builder", Object);
 
-    NODE_SET_PROTOTYPE_METHOD (constructor_template, "get", Get);
+    DEFINE_NODE_METHOD ("get", Get);
+
     target->Set (String::NewSymbol ("Builder"), t->GetFunction ());
 }
